+++ inst/examples/research-pipeline/analysis/pipeline_orchestration.R
@@ -1,0 +1,6 @@
+library(targets)
+library(plumber)
+
+.run_pipeline_stub <- function() {
+  invisible(NULL)
+}